requirements.txt
torch>=2.8.0+cu128
torchvision>=0.23.0+cu128
numpy
pandas
matplotlib
scikit-learn
seaborn
notebook
tqdm
Pillow
